refactor(single): extract executive.rs from single/mod.rs
Moves build_positive_signals, build_executive_narrative,
build_single_assessment_text, build_single_key_points_text,
build_single_quick_actions_text, and sentence_preview into
src/output/builder/single/executive.rs. No logic changes.
